Cambios en Sistema para sugerir Productos
Se crea el metodo sugerirProductos asi una vez que termina de sugerir los productos segun su preferencia se le muestran otros disponibles hasta no tener tiempo o plata, como pide la consigna.
Se arreglan algunas variables que estaban como int e iban como double en Sistema y Usuario.

+++ TurismoenlaTierraMedia/src/TurismoenlaTierraMedia/Sistema.java
@@ -93,7 +93,7 @@ public class Sistema {
 	// --------------------
 	// GRABAR COMPRAS
 
-	public static void escribirCompras(String nombre, double tiempoTotal, int dineroTotal, List<Atracciones> producto,
+	public static void escribirCompras(String nombre, double tiempoTotal, double dineroTotal, List<Atracciones> producto,
 			String file)
 
 			throws IOException {
@@ -121,7 +121,6 @@ public class Sistema {
 			ArrayList<String> nombresAtrIncluidas = compra.getNombreAtracciones();
 
 			for (int i = 0; i < nombresAtrIncluidas.size(); i++) {
-
 				salida.println(nombresAtrIncluidas.get(i));
 			}
